backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
opentelemetry-proto>=1.22.0
protobuf>=4.0
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
python-dotenv>=1.0.0
redis>=5.0.0
rq>=1.16.0
openai>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
